fix: send silence/blackness when tracks are muted (#47)
This fixes the timestamp freeze observed by backend components such as avp leading to lip sync issues. In future, the need of sending dummy tracks may be rendered redundant when we implement algorithms to use RTCP timestamps for audio video sync.

For audio track, setting enabled to false makes the browser send silence. We may still need to have hold on the mic resource to show prompt if the user is speaking.

For video track, setting enabled to false is not enough as the camera resource is not given up shown by the glowing camera light in a computer which may feel invasive. So a dummy track is created and sender.replacetrack is used to replace on demand.