feat: send blank tracks when video is muted (#105)
* feat: send blank tracks when video is muted

Aim - AV Sync issues
there is a hypothesis that no packets being sent when the video is muted might cause av sync. This PR ensures that summy packets are sent at 1fps after video mute.

Implementation note:
canvas.captureStream() even with the framerate option or with request frame doesn't trigger the packet send which only happens when there is a change in the canvas stream. To simulate the change the red value of a pixel on the canvas is flipped periodically.

* fix: rename test file so it doesn't create failure